Home, specific, App.js and Css files updated.
New code done in the homepage to get the venues on the page. Created and added new code in the specific page to get a specific venue with a id on the page. In the App.js added a new import and a new route path for the specific page. Also some new styles added to the css file.